requirements.txt
faust-streaming==0.10.8
pip-chill==1.0.1
requests==2.28.2
